+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ dev branch (upcoming version)
 - BO/BOLFI: take advantage of priors
 - BO/BOLFI: take advantage of seed
 - BO/BOLFI: improved optimization scheme
+- BO/BOLFI: bounds must be a dict
 
 
 0.5 (2017-05-19)